initial upload of hexagonal mirror script
Runs in Jupyter Notebook using PIL, Numpy, and STL. The script samples regions of the image to create averaged hexagon values, so the sample image width and height is ~10x larger than the intended number of mirrors. We should probably come up with a better way to create precise values for each hex if desired.